requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
PyPDF2
chromadb
sentence-transformers